Fix nodes command to display uptime as wallclock seconds

+++ apps/shell/jamshell.js
@@ -1,7 +1,7 @@
 jdata {
 	char* jobs as logger;
     struct device {
-        float uptime;
+        int uptime;
         char* nodeType;
     } info as logger;
 }
